One more cleanup
