+++ Deanon/Deanon/Configuration/Config.cs
@@ -1,0 +1,11 @@
+﻿namespace Deanon.Configuration
+{
+    public class Config
+    {
+        public Vk VK { get; set; }
+        public Db DB { get; set; }
+        public Depth Depth { get; set; }
+        public Stages Stages { get; set; }
+        public Exec Exec { get; set; }
+    }
+}
