+++ Deanon/Deanon/Configuration/Db.cs
@@ -1,0 +1,10 @@
+﻿namespace Deanon.Configuration
+{
+    public class Db
+    {
+        public string Address { get; set; }
+        public int Port { get; set; }
+        public string User { get; set; }
+        public string Password { get; set; }
+    }
+}
