+++ Deanon/Deanon/Configuration/Depth.cs
@@ -1,0 +1,11 @@
+﻿namespace Deanon.Configuration
+{
+    public class Depth
+    {
+        public int Friends { get; set; }
+        public int Followers { get; set; }
+        public int Post { get; set; }
+        public int Comments { get; set; }
+        public int Likes { get; set; }
+    }
+}
