+++ Deanon/Deanon/Configuration/Exec.cs
@@ -1,0 +1,10 @@
+﻿namespace Deanon.Configuration
+{
+    public class Exec
+    {
+        public int UserId { get; set; }
+        public bool ContinueFromSavePoint { get; set; }
+        public bool CompleteRelations { get; set; }
+        public int Expansions { get; set; }
+    }
+}
