+++ Deanon/Deanon/Configuration/Stages.cs
@@ -1,0 +1,8 @@
+﻿namespace Deanon.Configuration
+{
+    public class Stages
+    {
+        public bool Small { get; set; }
+        public bool Big { get; set; }
+    }
+}
