+++ Deanon/Deanon/Configuration/Vk.cs
@@ -1,0 +1,7 @@
+﻿namespace Deanon.Configuration
+{
+    public class Vk
+    {
+        public string Token { get; set; }
+    }
+}
