+++ Deanon/Deanon/DB/DataModels/Relationships/RelationData/InteractionData.cs
