+++ Deanon/Deanon/DB/DataModels/Relationships/RelationData/InteractionType.cs
