--- Deanon/Deanon/dumper/EnterType1.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Deanon.dumper
-{
-    public enum EnterType
-    {
-        Friend,
-        Post,
-        Comments,
-        Follower,
-        Likes
-    }
-}
